CPU Arch Done

+++ CPU/Bus.cpp
@@ -17,12 +17,12 @@ Bus::~Bus() {
 
 }
 
-void Bus::write(uint8_t addr, uint16_t data) {
+void Bus::write(uint16_t addr, uint8_t data) {
     if (addr >= 0x0000 && addr <= 0xFFFF)
         ram[addr] = data;
 }
 
-uint16_t Bus::read(uint16_t addr, bool bReadOnly) {
+uint8_t Bus::read(uint16_t addr, bool bReadOnly) {
     if (addr >= 0x0000 && addr <= 0xFFFF)
         return ram[addr];
